Correctly setup netfiler conntrack handle for CT dump.
Currently, the netlink handle used to take conntrack dump is
inadvertently set with subscriptions: NFCT_ALL_CT_GROUPS. This
makes kernel deliver all event messages in the dump handle as well.
This lead to leak of ipv6 based entries in the dump_callback and leading
to segmentation fault.

As per the libnetfilter_conntrack documentation/examples, instead the
subscriptions should be set to 0, so that kernel also provides the dump
related information and honour the family param set during nfct_query.

Signed-off-by: Priyankar Jain <priyankar.jain@nutanix.com>