conntrack_store: Make CT update event update processing atomic
There are 2 threads which are listening on CT events (one thread
on the src IP based events, and the second thread based on the
destination IP events).

There is a case in which both threads can receive the same conntrack
event (update or delete). (CT entry's src and dst IP are both part of
migratable IP given to the process.)

Now during processing of such an update following happens:

1. The current entry is looked up from the conntrack store.
2. A new entry is generated based on the current entry from CT store and
   the data we got from the event handler.
3. Finally the old/current entry is freed.

Step1, 2 and 3 are atomic amongst themselves but 1,2,3 is not an atomic
operation as a whole.

There is a possibility that wrt threads T1 T2, following sequence of
operations happens:

T1: 1, 2
T2: 1
T1: 3
T2: 2, 3

So when T2 performs 2, since the entry is already freed by T1 (at
step-3) segmentation fault happens when T2 accesses that memory at
its step-2.

This commit fixes it by making 1,2,3 an atomic operation as a whole.

Signed-off-by: Priyankar Jain <priyankar.jain@nutanix.com>