tests: Fix conntrack UTs by not relying on ct_mark for DELETE event. (#11)
For the purpose of testing we are using ct mark as id because
ct_id is not known while creating the entries, so validation becomes
easier. All new and update events created by test will have ct_mark
set.

For kernel version 5.x, we are not getting ct_mark for delete
event from kernel. For delete event instead of using ct_mark,
with this commit we are now iterating over the complete hashtable
and match entries based on ct_id.
Similarly, entries without ct_mark are ignored for NEW/UPDATE events.

NOTE: This change is required only for tests because it is
utilising ct_mark as a way for verification of entries. Actual dev code
does not rely on ct_mark instead uses ct_id for all CT operations.

Signed-off-by: Priyankar Jain <priyankar.jain@nutanix.com>