Don't insist on getting a DBus address via env
It doesn't seem to be actually used anywhere in the code? The daemon
code already does the sensible thing and just uses library functions to
say "connect to system bus". The only use case for such a setup I could
see would be testing, but there's no test harness that makes use of
this.

Change-Id: I1098fdc33980e2f065a09bd9c0b698b5f5977d53